fixup! feat: Server-Side Player Attributes, Client-Side Function Binds refactor and LUA Documentation
